+++ src/android/nl/xservices/plugins/accessor/CalendarProviderAccessor.java
@@ -99,7 +99,7 @@ public class CalendarProviderAccessor extends AbstractCalendarAccessor {
   public String createEvent(Uri eventsUri, String title, long startTime, long endTime,
                             String description, String location, Long firstReminderMinutes, Long secondReminderMinutes,
                             String recurrence, int recurrenceInterval, String recurrenceWeekstart,
-                            String recurrenceByDay, String recurrenceByMonthDay, Long recurrenceEndTime, Long recurrenceCount,
+                            String recurrenceByDay, String recurrenceByMonthDay, Long recurrenceEndTime, int recurrenceCount,
                             String allday, Integer calendarId, String url) {
     eventsUri = eventsUri == null ? Uri.parse(CONTENT_PROVIDER + CONTENT_PROVIDER_PATH_EVENTS) : eventsUri;
     return super.createEvent(eventsUri, title, startTime, endTime, description, location,
